# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Invoice API
  description: |
    Multi-tenant SaaS API for invoice management.

    ## Authentication
    All protected endpoints require `Authorization: Bearer <token>` header.

    Multi-company access uses `X-Company-Id` header to select the active company context.

    ## Roles
    - **SUPER_ADMIN** — Full access. Can create companies, add users, manage everything.
    - **ADMIN** — Can view and manage document types.
    - **USER** — Basic access within a company.

    ## Getting Started
    1. Register via `POST /auth/register`
    2. Verify email via `POST /auth/verify`
    3. Login via `POST /auth/login` to get JWT token
    4. Use token in `Authorization: Bearer <token>` header
    5. Create a company via `POST /companies`
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - Bearer Authentication: []
components:
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token obtained from login
  schemas:
...
    AddUserToCompanyRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
          description: User email address. If user does not exist, a new account is created.
          example: john@example.com
        firstName:
          type: string
          description: First name (required if user does not exist)
          example: John
        lastName:
          type: string
          description: Last name
          example: Doe
        username:
          type: string
          description: Username (required if user does not exist)
          example: johndoe
        password:
          type: string
          description: Password (required if user does not exist)
          example: Secret@123
        role:
          type: string
          description: Role to assign in the company
          enum:
            - SUPER_ADMIN
            - ADMIN
            - USER
          example: ADMIN
...
    CompanyResponse:
      type: object
      properties:
        companyId:
          type: string
          format: uuid
          description: Company UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Company name
          example: Acme Corp
        urlSlug:
          type: string
          description: URL slug for SEO-friendly URLs
          example: acme-corp
        description:
          type: string
          description: Company description
          example: We build amazing software
        addressLine1:
          type: string
          description: Address line 1
          example: 123 Business Park
        addressLine2:
          type: string
          description: Address line 2
          example: Building A, Floor 3
        city:
          type: string
          description: City
          example: Mumbai
        state:
          type: string
          description: State
          example: Maharashtra
        country:
          type: string
          description: Country
          example: India
        pinCode:
          type: string
          description: PIN code
          example: "400001"
        contactNumber:
          type: string
          description: Contact number
          example: "9876543210"
        hasLogo:
          type: boolean
          description: Whether company has a logo
          example: true
        logoUrl:
          type: string
          description: Logo URL
          example: /api/v1/companies/550e8400-e29b-41d4-a716-446655440000/logo
        createdBy:
          type: string
          format: uuid
          description: Created by user UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-01-01T00:00:00Z"
        active:
          type: boolean
          description: Whether company is active
          example: true
        userRole:
          type: string
          description: Current user's role in this company
          enum:
            - SUPER_ADMIN
            - ADMIN
            - USER
          example: SUPER_ADMIN
        message:
          type: string
          description: Status message
          example: Company created successfully
    CompanyUserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: User email
          example: john@example.com
        firstName:
          type: string
          description: First name
          example: John
        lastName:
          type: string
          description: Last name
          example: Doe
        role:
          type: string
          description: Role in the company
          enum:
            - SUPER_ADMIN
            - ADMIN
            - USER
          example: ADMIN
        active:
          type: boolean
          description: Whether the membership is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: When the user was added to the company
          example: "2026-01-01T00:00:00Z"
        message:
          type: string
          description: Status message
          example: User added to company
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2026-01-01T00:00:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Validation failed
        path:
          type: string
          description: Request path
          example: /api/v1/auth/register
        validationErrors:
          type: object
          description: Field-level validation errors (only for 400 validation errors)
          additionalProperties:
            type: string
          example:
            email: Email should be valid
            password: Password must be at least 8 characters
paths:
...
  /companies:
    post:
      tags:
        - Companies
      summary: Create a new company
      description: Creates a new company. The authenticated user is automatically assigned SUPER_ADMIN role.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - addressLine1
                - city
                - state
                - country
                - pinCode
                - contactNumber
              properties:
                name:
                  type: string
                  description: Company name (must be unique)
                  example: Acme Corp
                aboutCompany:
                  type: string
                  description: About the company
                  example: We build amazing software
                addressLine1:
                  type: string
                  description: Address line 1
                  example: 123 Business Park
                addressLine2:
                  type: string
                  description: Address line 2
                  example: Building A, Floor 3
                city:
                  type: string
                  description: City
                  example: Mumbai
                state:
                  type: string
                  description: State
                  example: Maharashtra
                country:
                  type: string
                  description: Country
                  example: India
                pinCode:
                  type: string
                  description: PIN code (6 digits)
                  example: "400001"
                contactNumber:
                  type: string
                  description: Contact number (10 digits)
                  example: "9876543210"
                logo:
                  type: string
                  format: binary
                  description: Company logo image (max 50KB)
      responses:
        '201':
          description: Company created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token not provided or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Company name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Companies
      summary: List user's companies
      description: Returns all companies the authenticated user has access to.
      responses:
        '200':
          description: Companies retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyResponse'
        '401':
          description: Token not provided or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /companies/{companyId}:
    get:
      tags:
        - Companies
      summary: Get company by ID
      description: Returns a specific company's details.
      parameters:
        - name: companyId
          in: path
          required: true
          description: Company UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '401':
          description: Token not provided or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied to this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Companies
      summary: Update company
      description: |
        Updates company details. Requires SUPER_ADMIN role.
        Company name cannot be changed. All fields are optional.
      parameters:
        - name: companyId
          in: path
          required: true
          description: Company UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                description:
                  type: string
                  description: Company description
                  example: Updated description
                addressLine1:
                  type: string
                  description: Address line 1
                  example: 456 New Avenue
                addressLine2:
                  type: string
                  description: Address line 2
                  example: Floor 5
                city:
                  type: string
                  description: City
                  example: Pune
                state:
                  type: string
                  description: State
                  example: Maharashtra
                country:
                  type: string
                  description: Country
                  example: India
                pinCode:
                  type: string
                  description: PIN code (6 digits)
                  example: "411001"
                contactNumber:
                  type: string
                  description: Contact number (10 digits)
                  example: "9876543211"
                logo:
                  type: string
                  format: binary
                  description: New company logo image (max 50KB)
      responses:
        '200':
          description: Company updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token not provided or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a SUPER_ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /companies/{companyId}/logo:
    get:
      tags:
        - Companies
      summary: Get company logo
      description: Returns the company logo image as binary data.
      parameters:
        - name: companyId
          in: path
          required: true
          description: Company UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Logo image (binary PNG)
          content:
            image/png:
              schema:
                type: string
                format: binary
        '401':
          description: Token not provided or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No logo attached to this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /companies/{companyId}/users:
    post:
      tags:
        - Companies
      summary: Add user to company
      description: |
        Adds a user to a company with a specific role.
        If the user does not exist in the system, a new account is created in PENDING_VERIFICATION
        status and a verification email is sent.
        Requires SUPER_ADMIN role.
      parameters:
        - name: companyId
          in: path
          required: true
          description: Company UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserToCompanyRequest'
      responses:
        '201':
          description: User added to company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyUserResponse'
        '400':
          description: Validation failed or missing required fields for new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token not provided or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a SUPER_ADMIN of this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Company not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User is already a member of this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Document Types ─────────────────────────────────────────
...